% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg.R
\name{agg}
\alias{agg}
\title{Tidy aggregate output}
\usage{
agg(...)
}
\arguments{
\item{...}{Parameters passed to \code{\link{aggregate}}.}
}
\value{
Data frame.
}
\description{
Tidy aggregate output
}
\details{
When calling multiple functions in a vector, \code{\link{aggregate}} nests the output into the dependent variable(s). To maintain a tidy data frame, \code{agg} replicates the procedure of \code{\link{aggregate}} and unnests the result.
}
\examples{
# Compare
ms   <- function(x) c(m = mean(x), s = sd(x))
form <- formula(cbind(mpg, disp) ~ am + gear)

A <- aggregate(form, mtcars, ms) # str(A) # Nested mean & sd in mpg and disp.
B <- agg(form, mtcars, ms)       # str(B) # Unnested.

}
\seealso{
\url{https://github.com/robertschnitman/afp}, \code{\link{aggregate}}, \code{\link{do.call}}, \code{\link{data.frame}}
}
