% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applytm.R
\name{applytm}
\alias{applytm}
\title{Multivariate tapply()}
\usage{
applytm(f = NULL, x, index, ...)
}
\arguments{
\item{f}{Function to apply. See \code{\link{tapply}}.}

\item{x}{Object where \code{\link{split}} is applicable. See \code{\link{tapply}}.}

\item{index}{List of indices. See \code{\link{tapply}} for univariate case.}

\item{...}{Parameters passed to \code{\link{mapply}}.}
}
\value{
List.
}
\description{
Apply a function over an array by a list of indices. Multivariate version of \code{\link{tapply}}.
}
\details{
Unlike \code{\link{tapply}}, one can pass a list of datasets and indices to \code{\link{applytm}}.
}
\examples{
# For each variable in a dataset, obtain means by a different index.
A <- mtcars[, c('mpg', 'wt', 'disp')]  # Targets.
B <- mtcars[, c('gear', 'am', 'carb')] # Indices.
applytm(A, B, mean)

}
\seealso{
\url{https://github.com/robertschnitman/afp}, \code{\link{tapply}}, \code{\link{mapply}}, \code{\link{dot}},
}
