% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcast.R
\name{bcast}
\alias{bcast}
\title{Compact mapply() a la` broadcast() from Julia}
\usage{
bcast(f, x, ..., simplify = FALSE)
.(f, x, ..., simplify = FALSE)
}
\arguments{
\item{f}{Function.}

\item{x}{First argument to vectorize over.}

\item{...}{a list of additional arguments to FUN. See \code{\link{mapply}}.}

\item{simplify}{Boolean, same as SIMPLIFY in \code{\link{mapply}}.}
}
\value{
List (if simplify = FALSE) or array (if simplify = TRUE).
}
\description{
Inspired by the dot syntax from Julia's \code{broadcast}, execute mapply() with a convenient shortand.
}
\examples{
set.seed(1)
a <- matrix(1:9, 3, 3)
b <- 20:22
c <- matrix(rnorm(9), 3,)
bcast(`/`, a, b, simplify = TRUE) # matrix.
bcast(`/`, a, b, c, simplify = TRUE) # If a warning occurs, output may be list.

# Alternate syntax
.(`/`, a, b, simplify = TRUE) # matrix.
.(`/`, a, b, c, simplify = TRUE) # If a warning occurs, output may be list.

}
\seealso{
\url{https://github.com/robertschnitman/afp}, \code{\link{mapply}},  
\code{broadcast} from Julia: \url{https://docs.julialang.org/en/v0.6.1/manual/arrays/#Broadcasting-1}
}
