% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.bind.R
\name{do.bind}
\alias{do.bind}
\title{Wrap do.call(*bind(), lapply()) into one function call}
\usage{
do.bind(f, x, rc = 1, ...)
}
\arguments{
\item{f}{A function to apply to the collection.}

\item{x}{A collection, such as a list, matrix, or dataframe.}

\item{...}{Arguments passed to \code{\link{do.call}}.}
}
\value{
Matrix or dataframe.
}
\description{
Wrap do.call(*bind(), lapply()) into one function call
}
\details{
After applying functions such as \code{\link{lapply}} or \code{\link{Map}} to a dataset, \code{\link{do.call}} is often a solution to combine the list elements to obtain the original a tabular format. 

To simplify this process, \code{\link{lapply}} and\code{\link{do.call}} are wrapped into this function, \code{do.bind}: it calls a function on a dataset and binds the elements into a matrix or dataframe (depending on the original collection type).

The type of bind is defined in a binary fashion (1 for row-wise, 2 for column-wise).
}
\examples{
# 1. Store lm() coefficients in matrix.
split1  <- split(mtcars, mtcars$gear)
adhoc1  <- function(s) {coef(lm(mpg ~ disp + wt + am, s))}
output1 <- do.bind(adhoc1, split1)
output1 # matrix. Rownames indicate subset.

# 2. What is the median Ozone-Temperature ratio for each given month?
airquality2 <- na.omit(airquality)
split2      <- split(airquality2, airquality2$Month)
adhoc2      <- function(x, y) {median(x/y)}
output2     <- do.bind(function(s) with(s, adhoc2(Ozone, Temp)), split2, 2)
output2 # matrix


}
\seealso{
\url{https://github.com/robertschnitman/afp}, \code{\link{lapply}}, \code{\link{do.call}}, \code{\link{rbind}}, \code{\link{cbind}}
}
