% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot.r
\name{dot}
\alias{dot}
\alias{.}
\title{Compact mapply()}
\usage{
.(...)
}
\arguments{
\item{...}{Same inputs as \code{\link{mapply}}.}
}
\value{
List.
}
\description{
Inspired by the dot syntax from Julia's \code{broadcast}, execute mapply() with a convenient shortand.
}
\seealso{
\url{https://github.com/robertschnitman/afp}, \code{\link{mapply}},  
\code{broadcast} from Julia: \url{https://docs.julialang.org/en/v0.6.1/manual/arrays/#Broadcasting-1}
}
