% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbind.R
\name{mapbind}
\alias{mapbind}
\title{Append two datasets' rows in a pairwise fashion}
\usage{
mapbind(x, y), mapbind_df(x, y)
}
\arguments{
\item{x}{2D object (e.g. Data frame, matrix)}

\item{y}{Same as above.}
}
\value{
Same as above; except in the case of mapbind_df, which outputs a data frame.
}
\description{
Append two datasets' rows in a pairwise fashion. In other words, append the 1st row of the first dataset with the 1st row of the second dataset, the 2nd row of the first with the 2nd row of the second, and so on.
}
\details{
This function can be useful for kable-friendly frequency distribution tables.
}
\examples{
library(tidyverse)
library(knitr)

freq_df <- with(diamonds, table(color, clarity)) \%>\%
  as.data.frame() \%>\%
  spread(NCOL(.) - 1, NCOL(.)) \%>\%
  mutate(Total = apply(.[, 2:NCOL(.)], 1, sum))

 prop_df <- freq_df
 prop_df[, 2:NCOL(prop_df)] <- prop_df[, 2:NCOL(prop_df)]/prop_df$Total
 prop_df[, 2:NCOL(prop_df)] \%<>\% 
   map_df(~ paste0(round(.x*100), '\%'))
  freq_df \%<>\% 
  format(., big.mark = ',', scientific = FALSE) \%>\%
  map_df(as.character)
  
prop_df \%<>\% map_df(as.character)

mapbind_df(freq_df, prop_df) \%>\% kable()

}
\seealso{
\url{https://github.com/robertschnitman/afp}, \code{\link{lapply}}, \code{\link{do.call}}, \code{\link{rbind}}, \url{https://purrr.tidyverse.org/}
}
