% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapchr.R
\name{mapchr}
\alias{mapchr}
\title{Apply functions to character vectors}
\usage{
mapchr(f, x)
reverse(x)
anagram(x)
}
\arguments{
\item{f}{Function to apply to the character vector `x`.}

\item{x}{Character vector.}
}
\value{
Vector.
}
\description{
Apply functions that manipulate character/string vectors.
}
\details{
`mapchr()` is a general functional for altering character vectors: apply any function to each of its element. The function `reverse()` reverses the order of characters for each element. `anagram(x)` randomly changes the order of the characters in every element.
}
\examples{
rn_mc <- rownames(mtcars)
mapchr(function(x) paste0(x, collapse = '|'), rn_mc)
reverse(rn_mc)
anagram(rn_mc)

}
