% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapdims.R
\name{mapdims}
\alias{mapdims}
\title{Map over matrix dimensions separately in a single function call}
\usage{
mapdims(f, x, ...)
mapc(f, x)
mapr(f, x)
}
\arguments{
\item{f}{A function to apply to the 2D dataset.}

\item{x}{A dataset, such as an array or data frame.}

\item{...}{Arguments passed to the \code{f}.}
}
\value{
List for mapdims(). Two elements: \code{rowwise} for the row-wise computations, and \code{colwise} for the column-specific results. Output is array for `mapc()` and `mapr()`
}
\description{
Map over matrix dimensions separately in a single function call
}
\details{
Apply function over dimensions and save the column- and row-wise results separately in a list with `mapdims()`; convenient `apply()` abbreviations with `mapc()` and `mapr()`.
}
\examples{
mapdims(median, mtcars)
mapdims(mean, airquality, na.rm = TRUE)
mapc(median, mtcars)
mapr(median, mtcars) 

}
\seealso{
\url{https://github.com/robertschnitman/afp}, \code{\link{apply}}
}
