% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapfilter.R
\name{mapfilter}
\alias{mapfilter}
\title{Apply function over collection and then filter it with a predicate function}
\usage{
mapfilter(f, p, x, ...)
}
\arguments{
\item{f}{A function to apply to the collection.}

\item{p}{Predicate function to filter the data. See \code{\link{Filter}}.}

\item{x}{A collection, such as a list, matrix, or data frame.}

\item{...}{Arguments passed to \code{\link{mapply}}.}
}
\description{
Apply function over collection and then filter it with a predicate function
}
\details{
Effectively, this function calls \code{mapply} and \code{Filter}.

When operating on lists, combining \code{mapfilter} with \code{\link{telecast}} is advised if the list elements are disparate.
}
\examples{
# 1. Obtain the squared even elements from a vector.
A       <- 1:50
is.even <- function(x) x \%\% 2 == 0
output1 <- mapfilter(function(x) x^2, is.even, A)
output1

# 2. For each dataset and its set of variables, find the cubed odd elements.
l       <- list(mc = mtcars, aq = airquality, lcs = LifeCycleSavings)
is.odd  <- function(x) x \%\% 2 == 1
mf.odd  <- function(x) mapfilter(function(y) y^3, is.odd, x)
output2 <- telecast(mf.odd, l)
output2 # list of lists.

}
\seealso{
\url{https://github.com/robertschnitman/afp}, \code{\link{Map}}, \code{\link{Filter}}, \code{\link{telecast}}
}
