% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapreduce.R
\name{mapreduce}
\alias{mapreduce}
\title{Apply function over collection and then iteratively reduce it a la` mapreduce() from Julia}
\usage{
mapreduce(f, o, x, ...)
}
\arguments{
\item{f}{A function to apply to the collection.}

\item{o}{A binary operator. Typically arithmetic operators, but can be another (binary) function.}

\item{x}{A collection, such as a a list, matrix, or data frame.}

\item{...}{Arguments passed to \code{\link{Reduce}}.}
}
\value{
Vector. Can differ based on the given inputs. For example, if accumulate = TRUE (via \code{\link{Reduce}}), a vector of length > 1 is returned; otherwise, a 1-element result.
}
\description{
Apply function over collection and then iteratively reduce it a la` mapreduce() from Julia
}
\details{
A simplified \code{mapreduce} from Julia.

Base R has functionals that output a list by way of \code{\link{lapply()}} and \code{\link{Map}} (among others), while \code{\link{Reduce}} diminishes given elements in a consecutive manner until a single result remains. To reduce the output of a mapping, the mentioned functions are required. 

In turn, \code{mapreduce} simplifies this process.
}
\examples{
# 1. Apply a function to 3 matrices and consecutively divide them.
matrixl <- list(A = matrix(c(1:9), 3, 3), B = matrix(10:18, 3, 3), C = matrix(19:27, 3, 3))
output1 <- mapreduce(function(x) x^2 + 1, `/`, matrixl) 
output1 # matrix 3x3

# 2. Split a dataset, apply a ratio to each subset, and then sum said ratios.
split2  <- split(mtcars, mtcars$gear)
seqrng  <- function(x) {seq(min(x), mean(x), mean(x) - min(x))}
adhoc   <- function(x, y) {seqrng(x)/seqrng(y)}
output2 <- mapreduce(function(s) with(s, adhoc(mpg, disp)), `+`, split2)
output2 # numeric, length 2


}
\seealso{
\url{https://github.com/robertschnitman/afp}, \code{\link{Map}}, \code{\link{Reduce}}
\code{mapreduce} from Julia:\url{https://docs.julialang.org/en/v0.6.1/stdlib/collections/#Base.mapreduce-NTuple{4,Any}}
}
