% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mop.R
\name{mop}
\alias{mop}
\title{Mop an array with multiple functions}
\usage{
mop(f, s, x, m, ...)
}
\arguments{
\item{f}{A function, typically a binary operator.}

\item{s}{A summary statistic function such as \code{\link{mean}}.}

\item{x}{An array.}

\item{m}{Margin. 1 for rows, 2 for columns.}

\item{...}{Arguments passed to \code{\link{sweep}}.}
}
\description{
Based on \code{\link{sweep}}, operate on an array by a summary statistic function.
}
\details{
Essentially, \code{mop} is a wrapper for \code{sweep(x, MARGIN, apply(...), FUN)}. Useful for indexing variables by their means, for example, so that the magnitude of a value relative to its average is known.
}
\examples{
mop(`/`, mean, mtcars, 2) # == sweep(mtcars, 2, apply(mtcars, 2, mean), `/`)
mop(`/`, median, mtcars, 2)

}
\seealso{
\url{https://github.com/robertschnitman/afp}, \code{\link{sweep}}
}
