% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtapply.R
\name{mtapply}
\alias{mtapply}
\title{Multivariate tapply()}
\usage{
mtapply(X, INDEX, FUN = NULL, ...)
}
\arguments{
\item{X}{Object where \code{\link{split}} is applicable. See \code{\link{tapply}}.}

\item{INDEX}{List of indices. See \code{\link{tapply}} for univariate case.}

\item{FUN}{Function to apply. See \code{\link{tapply}}.}

\item{...}{Parameters passed to \code{\link{mapply}}.}
}
\value{
List.
}
\description{
Apply a function over an array by a list of indices. Multivariate version of \code{\link{tapply}}.
}
\details{
Unlike \code{\link{tapply}}, one can pass a list of datasets and indices to \code{\link{mtapply}}.
}
\examples{
# For each variable in a dataset, obtain means by a different index.
A <- mtcars[, c('mpg', 'wt', 'disp')]  # Targets.
B <- mtcars[, c('gear', 'am', 'carb')] # Indices.
mtapply(A, B, mean)

}
\seealso{
\url{https://github.com/robertschnitman/afp}, \code{\link{tapply}}, \code{\link{mapply}}
}
