% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reducemap.R
\name{reducemap}
\alias{reducemap}
\title{Reduce a dataset by a margin and apply a function afterward}
\usage{
reducemap(o, m = 2, f, x, ...)
}
\arguments{
\item{o}{A binary operator. Typically arithmetic operators, but can be another (binary) function.}

\item{m}{Margin. 1 = row-wise; 2 = column-wise (default). See \code{\link{apply}}.}

\item{f}{A function to apply to the collection.}

\item{x}{A collection, such as a a list, matrix, or data frame.}

\item{...}{Arguments passed to \code{\link{Reduce}}.}
}
\description{
Reduce a dataset by a margin and apply a function afterward. Effectively, this function calls \code{\link{mapply}} and \code{\link{reducechop}}.
}
\details{
Inverse of \code{mapreduce}: reduce a dataset row-wise or column-wise, applying a function on the collection afterward.
}
\examples{
reducemap(`+`, 2, function(x) x^2, mtcars) # Add all values for each variable and then square the reduction.
reducemap(`-`, 1, function(x) x^2, mtcars, right = TRUE) # Subtracts iteratively starting from the rightmost row

}
\seealso{
\url{https://github.com/robertschnitman/afp}, \code{\link{mapply}}, \code{\link{Reduce}}, \code{\link{reducechop}},  
\code{mapreduce} from Julia:\url{https://docs.julialang.org/en/v0.6.1/stdlib/collections/#Base.mapreduce-NTuple{4,Any}}
}
