% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapchr.R
\name{split_apply}
\alias{split_apply}
\title{Manipulate individual elements in a character vector}
\usage{
mapchr(f, x)
jumble(x)
is.upper(x)
is.lower(x)
}
\arguments{
\item{f}{Function to apply to the character vector `x`.}

\item{x}{Character vector.}
}
\value{
Vector.
}
\description{
Apply functions that manipulate character/string vectors. Useful for reordering the characters in each vector element, for example.
}
\details{
`mapchr()` is a general functional for altering character vectors: apply any function to each of its elements. `jumble(x)` randomly changes the order of the characters in every element. Similar to `map_chr()` from the `purrr` library with the exception that the former only accepts character vectors as the data input. These functions are useful when manipulating the arrangement of the characters is desired. Finally, `is.upper()` and `is.lower()` test whether each element in a string vector is all uppercase or lowercase.
}
\examples{
# 1. Manipulating the letter ordering in each vector elemetn.
rn_mc <- rownames(mtcars)
mapchr(function(x) paste0(x, collapse = '|'), rn_mc)
jumble(rn_mc)

# 2. Testing for uppercase or lowercase.
chr <- c('TEST', 'test', 'tEsT')
is.upper(chr) # TRUE FALSE FALSE
is.lower(chr) # FALSE  TRUE FALSE

}
\seealso{
\url{https://github.com/robertschnitman/afp}, \code{\link{lapply}}
}
