% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapdims.R
\name{mapdims}
\alias{mapdims}
\title{Map over matrix dimensions separately in a single function call}
\usage{
mapdims(f, x, ...)
}
\arguments{
\item{f}{A function to apply to the 2D dataset.}

\item{x}{A 2D dataset, such as a matrix or data frame.}

\item{...}{Arguments passed to the \code{f}.}
}
\value{
List. Two elements: \code{rowwise} for the row-wise computations, and \code{colwise} for the column-specific results.
}
\description{
Apply function over dimensions and save the column- and row-wise results separately in a list.
}
\examples{
mapdims(median, mtcars)

mapdims(mean, airquality, na.rm = TRUE)

}
\seealso{
\url{https://github.com/robertschnitman/afp}, \code{\link{apply}}
}
