% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mop.R
\name{mop}
\alias{mop}
\title{Mop an array with multiple functions}
\usage{
mop(x, m, s, f, ...), smop(x, s, f, ...)
}
\arguments{
\item{x}{An array.}

\item{m}{Margin. 1 for rows, 2 for columns.}

\item{s}{A summary statistic function such as \code{\link{mean}}.}

\item{f}{A function to be "swept" or "mopped" out, typically a binary operator.}

\item{...}{Arguments passed to \code{\link{sweep}}.}
}
\description{
Based on \code{\link{sweep}}, operate on an array by a summary statistic function.
}
\details{
Essentially, \code{mop} is a wrapper for \code{sweep(x, MARGIN, apply(...), FUN)}. Useful for indexing variables by their means, for example, so that the magnitude of a value relative to its average is known. The function \code{smop} simplifies \code{mop} by operating only on columns.
}
\examples{
mop(mtcars, 2, mean, `/`) # == sweep(mtcars, 2, apply(mtcars, 2, mean), `/`)
mop(mtcars, 2, median, `/`)

}
\seealso{
\url{https://github.com/robertschnitman/afp}, \code{\link{sweep}}
}
