% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telecast.r
\name{telecast}
\alias{telecast}
\title{Apply common function over multiple objects independent of each other}
\usage{
telecast(f, l)
}
\arguments{
\item{f}{A function to apply to the collection(s).}

\item{l}{List of data objects.}
}
\value{
List.
}
\description{
Apply common function over multiple objects independent of each other
}
\details{
Inspired by \code{broadcast} from Julia. Essentially, \code{telecast} wraps \code{\link{mapply}} within \code{\link{lapply}}.

Base R functions such as \code{\link{Map}}/\code{\link{mapply}} apply functions pairwise when given multiple data objects. \code{telecast} fills the need to apply a common function against data objects exclusive to each other. 

Useful for storing disparate information into a single list.
}
\examples{
l <- list(mc = mtcars, aq = airquality, lcs = LifeCycleSavings) # Store to reference easily for the two examples below.

## 1. Extract means for each variable from 3 datasets.
mean.nr <- function(x) mean(x, na.rm = TRUE) # airquality has NA values.
output1 <- telecast(mean.nr, l) 
output1 # Compare: lapply(l, function(x) mapply(mean.nr, x))

## 2. Derive distinct iterative reductions along with Reduce().
red.div <- function(y) Reduce(`/`, y)
output2 <- telecast(red.div, l) 
output2 # Compare: lapply(l, function(x) mapply(red.div, x))

}
\seealso{
\url{https://github.com/robertschnitman/afp}, \code{\link{mapply}}, \code{\link{lapply}}
\code{broadcast} from Julia: https://docs.julialang.org/en/v0.6.1/manual/arrays/#Broadcasting-1
}
